Full refactor Driver API
Updating in one all file for each architecture.
This file will contains all driver for personalized application
